requirements.txt
streamlit
google-generativeai
sqlalchemy
pymysql